# upload/extracted/Attached-Assets-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Цифровой обходной лист — Digital Routing Sheet HR system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /termination-sheets:
    get:
      operationId: listTerminationSheets
      tags: [termination]
      summary: List termination sheets (filtered by status/branchId)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of termination sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TerminationSheetSummary"
    post:
      operationId: createTerminationSheet
      tags: [termination]
      summary: HR creates a termination sheet for an employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminationSheetInput"
      responses:
        "201":
          description: Termination sheet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminationSheetDetail"
  /termination-sheets/{id}:
    get:
      operationId: getTerminationSheet
      tags: [termination]
      summary: Get termination sheet with all steps
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Termination sheet with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminationSheetDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /termination-sheets/{id}/close:
    post:
      operationId: closeTerminationSheet
      tags: [termination]
      summary: HR closes a completed termination sheet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Termination sheet closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminationSheetDetail"
  /termination-sheets/{id}/restore:
    post:
      operationId: restoreTerminationSheet
      tags: [termination]
      summary: Admin restores a stopped/rejected termination sheet (within 1 hour of rejection)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Termination sheet restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminationSheetDetail"
        "403":
          description: Restore window expired (>1 hour) or not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Termination Steps ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ─────────────────────────────────────────────
...
    TerminationSheetInput:
      type: object
      required: [employeeFullName, branchId, positionId, terminationDate]
      properties:
        employeeFullName:
          type: string
        branchId:
          type: number
        positionId:
          type: number
        terminationDate:
          type: string
          format: date-time
    TerminationSheetSummary:
      type: object
      required: [id, employeeFullName, branchId, positionId, isDoctor, terminationDate, initiatorName, status, createdAt]
      properties:
        id:
          type: number
        employeeFullName:
          type: string
        branchId:
          type: number
        branchName:
          type: string
        positionId:
          type: number
        positionName:
          type: string
        isDoctor:
          type: boolean
        terminationDate:
          type: string
          format: date-time
        initiatorId:
          type: number
        initiatorName:
          type: string
        status:
          type: string
          enum: [in_progress, completed, rejected, stopped]
        statusToken:
          type: ["string", "null"]
          description: Only included for admin/hr/recruiter roles
        rejectedById:
          type: ["number", "null"]
        rejectedByName:
          type: ["string", "null"]
        rejectedAt:
          type: ["string", "null"]
          format: date-time
        rejectionReason:
          type: ["string", "null"]
        stoppedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TerminationSheetDetail:
      type: object
      required: [id, employeeFullName, branchId, positionId, isDoctor, terminationDate, initiatorName, status, statusToken, steps, createdAt]
      properties:
        id:
          type: number
        employeeFullName:
          type: string
        branchId:
          type: number
        branchName:
          type: string
        positionId:
          type: number
        positionName:
          type: string
        isDoctor:
          type: boolean
        terminationDate:
          type: string
          format: date-time
        initiatorId:
          type: number
        initiatorName:
          type: string
        status:
          type: string
          enum: [in_progress, completed, rejected, stopped]
        statusToken:
          type: string
        rejectedById:
          type: ["number", "null"]
        rejectedByName:
          type: ["string", "null"]
        rejectedAt:
          type: ["string", "null"]
          format: date-time
        rejectionReason:
          type: ["string", "null"]
        stoppedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TerminationStep"
        createdAt:
          type: string
          format: date-time

    # ── Termination Steps ─────────────────────────────────
    TerminationStep:
      type: object
      required: [id, terminationSheetId, stepType, assignedRole, status, isBlocking, createdAt]
      properties:
        id:
          type: number
        terminationSheetId:
          type: number
        stepType:
          type: string
          enum: [chief_physician_off, it_revocation, marketing_off, accounting_off, security_off, hr_exit_interview, hr_close, medical_equipment_off, account_manager_delete_profile]
        assignedRole:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, skipped]
        isBlocking:
          type: boolean
        comment:
          type: ["string", "null"]
        exitInterviewNotes:
          type: ["string", "null"]
        completedById:
          type: ["number", "null"]
        completedByName:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...